==> analysis/05-GetGBLUPs.Rmd <==
---
title: "Predict GEBV"
author: "wolfemd"
date: "2020-Sep-16"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = F, tidy = T)
```

# Previous step

4. [Check prediction accuracy](04-CrossValidation.html): Evaluate prediction accuracy with cross-validation.

# Objective

**Current Step**  

5. [Genomic prediction](05-GetGBLUPs.html): Predict _genomic_ BLUPs (GEBV and GETGV) for all selection candidates using all available data.

# Prediction using three-stages


## Set-up
```{bash, eval=F}
# activate multithread OpenBLAS 
export OMP_NUM_THREADS=1
```

```{r, eval=F}
library(tidyverse); library(magrittr); 
source(here::here("code","gsFunctions.R"))

A<-readRDS(file=here::here("output","Kinship_A_IITA_2020Sep16.rds"))
D<-readRDS(file=here::here("output","Kinship_D_IITA_2020Sep16.rds"))
AD<-readRDS(file=here::here("output","Kinship_AD_IITA_2020Sep16.rds"))

blups<-readRDS(file=here::here("output","iita_blupsForModelTraining.rds")) %>% 
  select(Trait,modelOutput) %>% 
  unnest(modelOutput) %>% 
  select(Trait,BLUPs) %>% 
  unnest(BLUPs) %>% 
  filter(GID %in% rownames(A)) %>% 
  nest(TrainingData=-Trait)
```

## Prediction

cbsurobbins (112 cores; 512GB) 

Model A
```{r, eval=F}
options(future.globals.maxSize= 1500*1024^2)
predModelA<-runGenomicPredictions(blups,modelType="A",grms=list(A=A),gid="GID",ncores=13)
saveRDS(predModelA,file = here::here("output","genomicPredictions_ModelA_threestage_IITA_2020Sep21.rds"))
```

Model ADE
```{r, eval=F}
options(future.globals.maxSize= 3000*1024^2)
predModelADE<-runGenomicPredictions(blups,modelType="ADE",grms=list(A=A,D=D,AD=AD),gid="GID",ncores=13)
saveRDS(predModelADE,file = here::here("output","genomicPredictions_ModelADE_threestage_IITA_2020Sep21.rds"))
```

## Write GEBVs
```{r, eval=F}
library(tidyverse); library(magrittr);
predModelA<-readRDS(file = here::here("output","genomicPredictions_ModelA_threestage_IITA_2020Sep21.rds"))
predModelADE<-readRDS(file = here::here("output","genomicPredictions_ModelADE_threestage_IITA_2020Sep21.rds"))
traits<-c("MCMDS","DM","PLTHT","BRNHT1","BRLVLS","HI","logFYLD","logTOPYLD","logRTNO","TCHART","LCHROMO","ACHROMO","BCHROMO")
```

```{r, eval=F}
predModelA %>% 
  dplyr::select(Trait,genomicPredOut) %>% 
  unnest(genomicPredOut) %>% 
  select(-varcomps) %>% 
  unnest(gblups) %>% 
  select(-GETGV) %>% 
  spread(Trait,GEBV) %>% 
  mutate(GeneticGroup=NA,
         GeneticGroup=ifelse(grepl("TMS18",GID,ignore.case = T),"TMS18",
                             ifelse(grepl("TMS15",GID,ignore.case = T),"TMS15",
                                    ifelse(grepl("TMS14",GID,ignore.case = T),"TMS14",
                                           ifelse(grepl("TMS13|2013_",GID,ignore.case = T),"TMS13","GGetc"))))) %>% 
  select(GeneticGroup,GID,all_of(traits)) %>% arrange(desc(GeneticGroup)) %>% 
  write.csv(., file = here::here("output","GEBV_IITA_ModelA_threestage_IITA_2020Sep21.csv"), row.names = F)
  
```
```{r, eval=F}
## Format and write GETGV
predModelADE %>% 
  dplyr::select(Trait,genomicPredOut) %>% 
  unnest(genomicPredOut) %>% 
  select(-varcomps) %>% 
  unnest(gblups) %>% 
  select(Trait,GID,GETGV) %>% 
  spread(Trait,GETGV) %>% 
  mutate(GeneticGroup=NA,
         GeneticGroup=ifelse(grepl("TMS18",GID,ignore.case = T),"TMS18",
                             ifelse(grepl("TMS15",GID,ignore.case = T),"TMS15",
                                    ifelse(grepl("TMS14",GID,ignore.case = T),"TMS14",
                                           ifelse(grepl("TMS13|2013_",GID,ignore.case = T),"TMS13","GGetc"))))) %>% 
  select(GeneticGroup,GID,all_of(traits)) %>% arrange(desc(GeneticGroup)) %>% 
  write.csv(., file = here::here("output","GETGV_IITA_ModelADE_threestage_IITA_2020Sep21.csv"), row.names = F)
```

```{r}
# gebv_vs_getgv<-predModelA %>% 
#   dplyr::select(Trait,genomicPredOut) %>% 
#   unnest(genomicPredOut) %>% 
#   select(-varcomps) %>% 
#   unnest(gblups) %>% 
#   select(-GETGV) %>% 
#   left_join(predModelADE %>% 
#               dplyr::select(Trait,genomicPredOut) %>% 
#               unnest(genomicPredOut) %>% 
#               select(-varcomps) %>% 
#               unnest(gblups) %>% 
#               select(Trait,GID,GETGV)) %>% 
#   mutate(GeneticGroup=NA,
#          GeneticGroup=ifelse(grepl("TMS18",GID,ignore.case = T),"TMS18",
#                              ifelse(grepl("TMS15",GID,ignore.case = T),"TMS15",
#                                     ifelse(grepl("TMS14",GID,ignore.case = T),"TMS14",
#                                            ifelse(grepl("TMS13|2013_",GID,ignore.case = T),"TMS13","GGetc")))))
```


```{r, fig.width=10, fig.height=10}
# gebv_vs_getgv %>% 
#   ggplot(.,aes(x=GEBV,y=GETGV,color=GeneticGroup)) + 
#   geom_point(alpha=0.7) + 
#   geom_abline(slope=1, color='darkred', linetype='dashed') + 
#   theme_bw() + 
#   facet_wrap(~Trait, ncol=3, scales='free') + 
#   scale_color_viridis_d()
```

# Prediction using two-stages

## Set-up
```{bash, eval=F}
# activate multithread OpenBLAS 
export OMP_NUM_THREADS=1
```

```{r, eval=F}
library(tidyverse); library(magrittr); 
source(here::here("code","gsFunctions.R"))

A<-readRDS(file=here::here("output","Kinship_A_IITA_2020Sep16.rds"))
D<-readRDS(file=here::here("output","Kinship_D_IITA_2020Sep16.rds"))
AD<-readRDS(file=here::here("output","Kinship_AD_IITA_2020Sep16.rds"))

# BLUPs from the 2 stage procedure
# (stage 1 of 2) using the 2019 procedure
blups<-readRDS(file=here::here("output","iita_blupsForModelTraining_twostage_asreml.rds")) %>% 
  dplyr::select(Trait,blups) %>% 
  unnest(blups) %>% 
  filter(GID %in% rownames(A)) %>% 
  nest(TrainingData=-Trait)
```

## Prediction

cbsurobbins (112 cores; 512GB) 

Model A
```{r, eval=F}
options(future.globals.maxSize= 1500*1024^2)
predModelA<-runGenomicPredictions(blups,modelType="A",grms=list(A=A),gid="GID",ncores=13)
saveRDS(predModelA,file = here::here("output","genomicPredictions_ModelA_twostage_IITA_2020Sep21.rds"))
```

Model ADE
```{r, eval=F}
options(future.globals.maxSize= 3000*1024^2)
predModelADE<-runGenomicPredictions(blups,modelType="ADE",grms=list(A=A,D=D,AD=AD),gid="GID",ncores=13)
saveRDS(predModelADE,file = here::here("output","genomicPredictions_ModelADE_twostage_IITA_2020Sep21.rds"))
```

## Write GEBVs
```{r}
library(tidyverse); library(magrittr);
predModelA<-readRDS(file = here::here("output","genomicPredictions_ModelA_twostage_IITA_2020Sep21.rds"))
predModelADE<-readRDS(file = here::here("output","genomicPredictions_ModelADE_twostage_IITA_2020Sep21.rds"))
traits<-c("MCMDS","DM","PLTHT","BRNHT1","BRLVLS","HI","logFYLD","logTOPYLD","logRTNO","TCHART","LCHROMO","ACHROMO","BCHROMO")
```

```{r}
predModelA %>% 
  dplyr::select(Trait,genomicPredOut) %>% 
  unnest(genomicPredOut) %>% 
  select(-varcomps) %>% 
  unnest(gblups) %>% 
  select(-GETGV,-contains("PEV")) %>% 
  spread(Trait,GEBV) %>% 
  mutate(GeneticGroup=NA,
         GeneticGroup=ifelse(grepl("TMS18",GID,ignore.case = T),"TMS18",
                             ifelse(grepl("TMS15",GID,ignore.case = T),"TMS15",
                                    ifelse(grepl("TMS14",GID,ignore.case = T),"TMS14",
                                           ifelse(grepl("TMS13|2013_",GID,ignore.case = T),"TMS13","GGetc"))))) %>% 
  select(GeneticGroup,GID,any_of(traits)) %>% arrange(desc(GeneticGroup)) %>% 
  write.csv(., file = here::here("output","GEBV_IITA_ModelA_twostage_IITA_2020Sep21.csv"), row.names = F)
  
```
```{r}
## Format and write GETGV
predModelADE %>% 
  dplyr::select(Trait,genomicPredOut) %>% 
  unnest(genomicPredOut) %>% 
  select(-varcomps) %>% 
  unnest(gblups) %>% 
  select(Trait,GID,GETGV) %>% 
  spread(Trait,GETGV) %>% 
  mutate(GeneticGroup=NA,
         GeneticGroup=ifelse(grepl("TMS18",GID,ignore.case = T),"TMS18",
                             ifelse(grepl("TMS15",GID,ignore.case = T),"TMS15",
                                    ifelse(grepl("TMS14",GID,ignore.case = T),"TMS14",
                                           ifelse(grepl("TMS13|2013_",GID,ignore.case = T),"TMS13","GGetc"))))) %>% 
  select(GeneticGroup,GID,any_of(traits)) %>% arrange(desc(GeneticGroup)) %>% 
  write.csv(., file = here::here("output","GETGV_IITA_ModelADE_twostage_IITA_2020Sep21.csv"), row.names = F)
```

```{r}
gebv_vs_getgv<-predModelA %>% 
  dplyr::select(Trait,genomicPredOut) %>% 
  unnest(genomicPredOut) %>% 
  select(-varcomps) %>% 
  unnest(gblups) %>% 
  select(-GETGV) %>% 
  left_join(predModelADE %>% 
              dplyr::select(Trait,genomicPredOut) %>% 
              unnest(genomicPredOut) %>% 
              select(-varcomps) %>% 
              unnest(gblups) %>% 
              select(Trait,GID,GETGV)) %>% 
  mutate(GeneticGroup=NA,
         GeneticGroup=ifelse(grepl("TMS18",GID,ignore.case = T),"TMS18",
                             ifelse(grepl("TMS15",GID,ignore.case = T),"TMS15",
                                    ifelse(grepl("TMS14",GID,ignore.case = T),"TMS14",
                                           ifelse(grepl("TMS13|2013_",GID,ignore.case = T),"TMS13","GGetc")))))
```


```{r, fig.width=10, fig.height=10}
gebv_vs_getgv %>% 
  ggplot(.,aes(x=GEBV,y=GETGV,color=GeneticGroup)) + 
  geom_point(alpha=0.7) + 
  geom_abline(slope=1, color='darkred', linetype='dashed') + 
  theme_bw() + 
  facet_wrap(~Trait, ncol=3, scales='free') + 
  scale_color_viridis_d()
```


# Next step

6. [Results](06-Results.html)
